requirements.txt
Babel==2.5.1
dj-database-url==0.4.2
Django==1.11.6
django-appconf==1.0.2
django-bootstrap-form==3.3
django-cors-headers==2.1.0
django-material==1.1.0
django-phonenumber-field==1.3.0
django-picklefield==1.0.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
gunicorn==19.7.1
phonenumberslite==8.8.4
psycopg2==2.7.3.2
pytz==2017.3
whitenoise==3.3.0
